+++ CHANGELOG.md
@@ -57,3 +57,4 @@
 - Enh #53: Introduce `Root` class for `root-level` HTML rendering and related tests (@terabytesoftw)
 - Bug #54: Update PHPDoc for `Voids` class for standard docs compliance (@terabytesoftw)
 - Enh #55: Introduce `Lists` class for `list-level` HTML rendering and related tests (@terabytesoftw)
+- Enh #56: Introduce `Table` class for `table-level` HTML rendering and related tests (@terabytesoftw)